rename component folder into _components
